fix(RandomNumbersGenerator) : rename variables
